servers/nina/pyproject.toml
[project]
name = "nina"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13, <4"
dependencies = [
    "fastmcp>=2.11.3",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
]
